Remove the word Clip
This project copies the filepath for whatever object is allowed using
the scoping functions: scope_media_panel_objects() or
scope_timeline_objects().

These scoping functions allow objects other than just PyClips, so it
doesn't make sense to have the word Clip in the name of this project.
For example, PySegments are allowed.

Remove the word Clip to indicate a broader scope of objects allowed.